Split AccountClient into partials
