--- tests/PinguApps.Appwrite.Client.Tests/Clients/AccountTests.cs
@@ -1,135 +1,0 @@
-﻿using System.Net;
-using Microsoft.Extensions.DependencyInjection;
-using PinguApps.Appwrite.Shared.Requests;
-using PinguApps.Appwrite.Shared.Tests;
-using Refit;
-using RichardSzalay.MockHttp;
-
-namespace PinguApps.Appwrite.Client.Tests.Clients;
-public class AccountTests
-{
-    private readonly MockHttpMessageHandler _mockHttp;
-    private readonly IAppwriteClient _appwriteClient;
-
-    public AccountTests()
-    {
-        _mockHttp = new MockHttpMessageHandler();
-        var services = new ServiceCollection();
-
-        services.AddAppwriteClientForServer("PROJECT_ID", Constants.Endpoint, new RefitSettings
-        {
-            HttpMessageHandlerFactory = () => _mockHttp
-        });
-
-        var serviceProvider = services.BuildServiceProvider();
-
-        _appwriteClient = serviceProvider.GetRequiredService<IAppwriteClient>();
-    }
-
-    [Fact]
-    public async Task Get_ShouldReturnSuccess_WhenApiCallSucceeds()
-    {
-        // Arrange
-        _mockHttp.Expect(HttpMethod.Get, $"{Constants.Endpoint}/account")
-            .ExpectedHeaders(true)
-            .Respond(Constants.AppJson, Constants.UserResponse);
-
-        _appwriteClient.SetSession(Constants.Session);
-
-        // Act
-        var result = await _appwriteClient.Account.Get();
-
-        // Assert
-        Assert.True(result.Success);
-    }
-
-    [Fact]
-    public async Task Get_ShouldHandleException_WhenApiCallFails()
-    {
-        // Arrange
-        _mockHttp.Expect(HttpMethod.Get, $"{Constants.Endpoint}/account")
-            .ExpectedHeaders(true)
-            .Respond(HttpStatusCode.BadRequest, Constants.AppJson, Constants.AppwriteError);
-
-        _appwriteClient.SetSession(Constants.Session);
-
-        // Act
-        var result = await _appwriteClient.Account.Get();
-
-        // Assert
-        Assert.True(result.IsError);
-        Assert.True(result.IsAppwriteError);
-    }
-
-    [Fact]
-    public async Task Create_ShouldReturnSuccess_WhenApiCallSucceeds()
-    {
-        // Arrange
-        var request = new CreateAccountRequest()
-        {
-            Email = "email@example.com",
-            Password = "password",
-            Name = "name"
-        };
-
-        _mockHttp.Expect(HttpMethod.Post, $"{Constants.Endpoint}/account")
-            .ExpectedHeaders()
-            .WithJsonContent(request)
-            .Respond(Constants.AppJson, Constants.UserResponse);
-
-        // Act
-        var result = await _appwriteClient.Account.Create(request);
-
-        // Assert
-        Assert.True(result.Success);
-    }
-
-    [Fact]
-    public async Task Create_ShouldHandleException_WhenApiCallFails()
-    {
-        // Arrange
-        var request = new CreateAccountRequest()
-        {
-            Email = "email@example.com",
-            Password = "password",
-            Name = "name"
-        };
-
-        _mockHttp.Expect(HttpMethod.Post, $"{Constants.Endpoint}/account")
-            .ExpectedHeaders()
-            .WithJsonContent(request)
-            .Respond(HttpStatusCode.BadRequest, Constants.AppJson, Constants.AppwriteError);
-
-        // Act
-        var result = await _appwriteClient.Account.Create(request);
-
-        // Assert
-        Assert.True(result.IsError);
-        Assert.True(result.IsAppwriteError);
-    }
-}
-
-public static class AccountTestsExtensions
-{
-    public static MockedRequest ExpectedHeaders(this MockedRequest request, bool addSessionHeaders = false)
-    {
-        var req = request
-            .WithHeaders("x-appwrite-project", Constants.ProjectId)
-            .WithHeaders("x-sdk-name", Constants.SdkName)
-            .WithHeaders("x-sdk-platform", "client")
-            .WithHeaders("x-sdk-language", Constants.SdkLanguage)
-            .WithHeaders("x-sdk-version", Constants.SdkVersion)
-            .WithHeaders("x-appwrite-response-format", Constants.AppwriteResponseFormat);
-
-        if (addSessionHeaders)
-            return req.ExpectSessionHeaders();
-
-        return req;
-    }
-
-    public static MockedRequest ExpectSessionHeaders(this MockedRequest request)
-    {
-        return request
-            .WithHeaders("x-appwrite-session", Constants.Session);
-    }
-}
